Stop characters from moving on a door tile
Characters only step on a door tile if the tile is part of a path. If
the door tile is the last tile in the path a walk action won't be
created.

If the player clicks the door tile directly we can be sure that he just
wants to open it (and he still can move there manually if needed). If
the player generates a path that leads through a door he probably
doesn't want to bother with opening it manually.