Add basics for calculating inaccurate shots
Characters get an accuracy skill which is used to calculate a random
chance to hit. We roll a random number between [1, 100] and if this
number is higher than the CTH the shot is accurate and lands right on
the clicked target. If the roll is smaller than the CTH the shot is
inaccurate and a random angle is applied to the shot (the worse the CTH
the bigger the range from which we choose the deviation of the shot).