Improve hit calculation
Instead of using magical dice rolls we use angles to determine the
derivation of a shot. Currently this is only influenced by the char's
marksman skill as well as his weapon's accuracy.

Those stats determine a maximum angle of derivation which will be used
to determine the flight path of the projectile.