Add simple bullet penetration / perforation
Each projectile has an energy level expressed as a percentage which is
reduced as soon as it hits an object or character on the map.

Each time the projectile hits an object the energy level will be
reduced and damage will be dealt to the object. If the energy level
reaches zero, the bullet will get stuck in the object.