Improve handling of destroyed world objects
Instead of letting the Tile class handle the removal / replacement of
destroyed world objects we moved the code to the Map class.

The map class already was updated every frame to reset the visibility
of seen tiles. The update function now also takes care of checking for
destroyed world objects and replacing them.

This has the benefit that the Tile class doesn't have an dependency on
the world object factory.