Use interaction cost when calculating a path
WorldObjects which can be interacted with like doors or fences have an
interaction cost. This will be used to determine how many AP it costs
to for example open this object.

The same cost will be used in the PathFinder to make sure we find the
shortest path in terms of AP usage.