Rewrote the way the game calculates FOV
Instead of doing so each frame update now only occur when a specific
action happens, such as a tile getting destroyed or a character
performs some action.

This helped to bring down the delta time a lot as the profiler shows:

Old | TOTAL TIME = 1.025436
New | TOTAL TIME = 0.004322

The code is a bit unstructured now and it still could slow down when
a lot of tiles are changing at the same time.

I decided against using the Messenger to notify the game about tile
changes because it feels a bit strange to "shout" around at the top
level where everybody can hear it.

That's why I turned the map into an observable and made the game listen
for events from it. I guess I'll implement more of the observer pattern.