Rewrite code for moving projectiles across map
Projectiles were moving across the map dependent on the game's time
step which made them miss tiles if they were too fast or during FPS
drops.

The current approach casts a ray from the origin to the target and gets
a list of all tiles in the path. The projectiles moves along those
tiles now instead of flying "freely".

This fixes a really big issue in the game which made hit calculations
inaccurate and led to bullets passing through walls and other
obstacles.

The second major issue it adresses is the separation of the mechanics
of the projectiles in the game and the animation layer.

The drawback is that animations don't look as nice and smooth anymore,
but this can be fixed later on.