Rewrite the inventory system
Instead of storage slots we now use inventory lists where items can be
added freely.

The new system also made it possible to add drag and drop functionality
to the InventoryScreen. To achieve this the different inventory items
are encapsulated in UIItems which can be dragged around between the
different lists.

The removal of storage slot also had a great benefit on the memory
usage. Before the change it was around 140mb and after the change it
dropped to about 26mb.