Get rid of Inventory class
Instead the Storage class has been extended to work with more specific
slots used for the player's equipment.

This breaks a few of the current inventory mechanics (e.g. trying to
equip an item into a slot which already contains an item). But the
current inventory system has a lot of flaws anyway and needs to be
replaced completely eventually.