Fix issues with behavior tree updates
The game makes sure that the behavior tree is left when no action can
be taken by the character. This gets rid of an issue where the game
would get stuck.

The other issue this fixes is the ticking of the behavior tree for
dead characters which lead to crashes and other problems.

Fixes #106.
Fixes #107.