Improve projectile handling code
Broke down the code into several smaller functions for better
readability and checking of the actual flow of the code.

For example the loop stopped early due to the returns even if multiple
projectiles still were on the map. It also looks like tiles were hit
twice if the projectile had 0 energy when hitting them.