Add basic procedural generation
The procedural generation works with a parcel system. First, it
generates a rough grid (for example 16x8 parcels). Each parcel
represents 8x8 tiles.

Then it tries to place prefabs on said parcel grid while applying
random rotation to add some variety. It also makes sure that none of
the prefabs are spawned next to each other (there has to be a distance
of at least one parcel between them) and none of them is directly at
the map border.

Closes #166.