Rewrite inventory user interface
Changed the layout to use a fixed screen size which allows for nicer
code and prevents rounding errors. Combined the item stats are with the
item description area and placed them below the equipment list.

Added proper scrollbars for the scrollareas even though they still seem
to have issues in some cases.

Some more minor changes to the functionality such as nicer highlights.

Closes #135, #170, #175.