Fix infinite loading bug on Mac OS 10.13
The error is caused by a bug in PhysFS, which doesn't report EOL
correctly to love.filesystem.lines and therefore causes the game to
hang endlessly while loading a file.

Added a quick hacky workaround to get rid of love.filesystem.lines for
now. Should be removed, once LÖVE implements the fixes.