Improve search for valid spawnpoints
Instead of randomly picking a coordinate on the map and checking it
against all spawnpoints, we now pick a random spawnpoint from the list
and check if it is valid.

Now it doesn't take more than 3 tries on average while before, it would
take up to a few hundred tries on larger maps.