Publish message queue events through tiles
Turned tiles into Observables which are observed by the Map object.
Game events can be published through the tiles now and will be
forwarded to the MessageQueue by the CombatState object.

This decouples the MessageQueue from the specific classes and also will
allow us to filter events which aren't visible to the player in future
updates.