Rewrite SoundManager
The SoundManager now loads and stores sounds according to their ids
defined in a separate file called info.lua.

It also is called directly now instead of observing the global
Messenger "channel". While this directly couples a lot of the games
modules to the SoundManager it is easier to understand what is
happening inside of the code than just seeing a Messenger call.

Closes #237.