Fix crash when switching to a new page
The crash was caused because the list tried to focus an UIElement at
the same cursor position. If the previous page contained (for example)
nine entries, it also tried to focus the ninth entry on the next page.

We now clamp the cursor to the max amount of items on the page. This
means the list will automatically focus the last item in the list if
there aren't enough items.