Remove separate list used in UIEquipmentList
The list was used to keep track of the UIEquipmentSlots, but they were
also added to the UIEquipmentList as children. When the list was
refreshed only the list was cleared, but slots were still added as
children. This caused the children table to grow continuously over
time.

Since the separate list was superfluous anyway we now use the children
table provided by the UIElement parent class to populate the equipment
list.

Fixes #246.