Remove separate list used in UIInventoryList
The list was used to keep track of the UIInventoryItems, which were
also added to the UIInventoryList as UIElement children. When the list
was refreshed the children weren't cleared. This caused the children
table to grow continuously over time.

Since the separate list was superfluous anyway we now use the children
table provided by the UIElement parent class to populate the inventory
list.

Fixes #247.