Hide tile info for unseen characters
The new function for checking wether a tile is visible which was
introduced in e7253886f08d1109f572c05e42d768bdd6f11931 allows us to
efficiently check wether a tile is visible to the player.

In this case we can now hide the info about enemy characters which
can't be seen by the player's characters.

Closes #255.