Improve mouse draggin behavior
Instead of hardcoding the dragging detection inside of InventoryScreen
we now check if the mouse is pressed and if the mouse has moved.

We then fire a "drag started" event, which can be received via the
appropriate ScreenManager methods.

Once the mouse is released we fire a "drag stopped" event.

Note: The mousepressed event is not suppressed!