Use percent based hit chance
The ProjectilePath checks via simple dice role wether the attack should
hit the target or miss. If it hits it determines all the tiles from
origin to target which need to be traversed.

If the shot misses we calculate a deviated target and return the
respective tiles to traverse.

The chance is only determined by the character's stats.

Closes #301.