Use cover values to calculate a CTH penalty
The cover system works by checking in which direction the attacker is
located in relation to the target (NSEW). We then check the tiles
facing in those directions. If they contain a world object which
provides a cover bonus we impose a "cover penalty" which is subtracted
from the original chance to hit.

Closes #299.